fix: match package and module structure with Hist
