+++ docs/api/cuda_histogram.axis.rst
@@ -1,0 +1,7 @@
+cuda_histogram.axis package
+===========================
+
+.. automodule:: cuda_histogram.axis
+   :members:
+   :undoc-members:
+   :show-inheritance:
