+++ docs/api/cuda_histogram.hist.rst
@@ -1,0 +1,7 @@
+cuda_histogram.hist module
+==========================
+
+.. automodule:: cuda_histogram.hist
+   :members:
+   :undoc-members:
+   :show-inheritance:
