--- docs/api/cuda_histogram.hist_tools.rst
@@ -1,8 +1,0 @@
-cuda_histogram.hist_tools module
-================================
-
-.. automodule:: cuda_histogram.hist_tools
-   :members:
-   :undoc-members:
-   :show-inheritance:
-   :private-members:
