+++ docs/api/cuda_histogram.rst
@@ -1,10 +1,18 @@
 cuda_histogram package
 ======================
 
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 3
+
+   cuda_histogram.axis
+
 Submodules
 ----------
 
 .. toctree::
    :maxdepth: 3
 
-   cuda_histogram.hist_tools
+   cuda_histogram.hist
